Re-organisation of the structure, use more of EHY tools generic functions.
svn path=/trunk/matlab/; revision=15174


Former-commit-id: 1063bf030f103087047e641f9c16a738e4353382

+++ applications/delft3d/nest_matlab/nesthd_check.m
@@ -4,28 +4,43 @@ function nesthd_check
 %
 %See also: nest_matlab
 %% Settings
-testdir         = 'c:\checkouts_svn\open_earth_repository\matlab\applications\delft3d\nest_matlab\check\';
-
-tests           = {[testdir filesep 'delft3d\milford']                       ; ...
-                   [testdir filesep 'delft3d\shenzen']                       ; ...
-                   [testdir filesep 'delft3d\zeedelta']                      ; };
+[path,~,~]      = fileparts(mfilename('fullpath'));
+testdir         = [path filesep 'check'];
 
 check_exe       = false;
 check_comprompt = true;
+check_nesthd1   = false;
+check_nesthd2   = true;
 
 %% set paths
-if ~isdeployed && any(which('setproperty'))
-   addpath(genpath('..\..\..\..\matlab'));
-end
+if ~isdeployed && any(which('setproperty')) addpath(genpath('..\..\..\..\matlab')); end
+
+%% Create list of directories
+tmp             = dir2(testdir);
+index           = [tmp.isdir];
+dirs            = tmp(index);
+for i_dir = 2: length(dirs) tests{i_dir - 1} = [testdir filesep dirs(i_dir).name]; end
 
 %% Create ini files
 for i_test = 1: length(tests)
-    copyfile([tests{i_test} filesep 'nesthd1.template'],[tests{i_test} filesep 'nesthd1.ini']);
-    copyfile([tests{i_test} filesep 'nesthd2.template'],[tests{i_test} filesep 'nesthd2.ini']);
-    substitute ('**rundir**',tests{i_test},[tests{i_test} filesep 'nesthd1.ini']);
-    substitute ('**rundir**',tests{i_test},[tests{i_test} filesep 'nesthd2.ini']);
+    nesthd1(i_test) = false;
+    nesthd2(i_test) = false;
+    if check_nesthd1
+        if exist([tests{i_test} filesep 'nesthd1.template'],'file')
+            nesthd1(i_test) = true;
+            copyfile([tests{i_test} filesep 'nesthd1.template'],[tests{i_test} filesep 'nesthd1.ini']);
+            substitute ('**rundir**',tests{i_test},[tests{i_test} filesep 'nesthd1.ini']);
+        end
+    end
+    
+    if check_nesthd2
+        if exist([tests{i_test} filesep 'nesthd2.template'],'file')
+            nesthd2(i_test) = true;
+            copyfile([tests{i_test} filesep 'nesthd2.template'],[tests{i_test} filesep 'nesthd2.ini']);
+            substitute ('**rundir**',tests{i_test},[tests{i_test} filesep 'nesthd2.ini']);
+        end
+    end
 end
-
 
 %% Windows batch file (first create)
 if check_exe
@@ -33,8 +48,8 @@ if check_exe
     for i_test = 1: length(tests)
         str_hd1 = ['nesthd ' tests{i_test} filesep 'nesthd1.ini'];
         str_hd2 = ['nesthd ' tests{i_test} filesep 'nesthd2.ini'];
-        fprintf(fid,'%s \n',str_hd1);
-        fprintf(fid,'%s \n',str_hd2);
+        if nesthd1(i_test) fprintf(fid,'%s \n',str_hd1); end
+        if nesthd2(i_test) fprintf(fid,'%s \n',str_hd2); end
     end
     fclose(fid);
 
@@ -42,18 +57,22 @@ if check_exe
     system('run.bat');
 
     for itest = 1: length(tests)
-        nesthd_compare([tests{itest} filesep 'nesthd2.ini']);
+        if nesthd2(i_test) nesthd_compare([tests{itest} filesep 'nesthd2.ini']); end
     end
 end
 
 %% Check if running from matlab command prompt works
 if check_comprompt
     for i_test = 1: length (tests)
-        nesthd        ([tests{i_test} filesep 'nesthd1.ini']);
-        nesthd_compare([tests{i_test} filesep 'nesthd1.ini']);
-        if i_test ~= 3
+        if nesthd1(i_test)
+            nesthd        ([tests{i_test} filesep 'nesthd1.ini']);
+            nesthd_compare([tests{i_test} filesep 'nesthd1.ini']);
+            delete        ([tests{i_test} filesep 'nesthd1.ini']); 
+        end
+        if nesthd2(i_test)
             nesthd        ([tests{i_test} filesep 'nesthd2.ini'],'check',true);
             nesthd_compare([tests{i_test} filesep 'nesthd2.ini']);
+            delete        ([tests{i_test} filesep 'nesthd2.ini']);
         end
     end
 end
